working copy 1

+++ Perceptron/Perceptron.h
@@ -15,7 +15,7 @@
 @property(strong, nonatomic) NSMutableArray* dataSet2;
 @property(strong, nonatomic) NSMutableArray* dataSet3;
 @property(strong, nonatomic) NSMutableArray* w;
-@property(assign, nonatomic) int learningRate;
+@property(assign, nonatomic) NSInteger learningRate;
 
 - (id)initWithHeader:(NSMutableArray*)header DataSetOne:(NSMutableArray*)one Two:(NSMutableArray*)two Three:(NSMutableArray*) three;
 - (void)doTraining;
